requirements.txt
# Z-Beam Generator - Python Dependencies

# Core dependencies
requests>=2.31.0
python-dotenv>=1.0.0
pydantic>=2.5.0

# Optional: For enhanced logging
rich>=13.0.0

# Development dependencies
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
